+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## [Unreleased]
 
 ### Added
+* Change order in structure view to match source file and sectioning level 
+* Add command redefinitions to command definition filter in structure view
 
 ### Fixed